backend/requirements.txt
Django == 3.2
django-mptt == 0.14.0
django-mysql == 4.7.1
djangorestframework == 3.14.0
djangorestframework-simplejwt == 5.2.2
Pillow == 9.0.0
python-dotenv
django-cors-headers  == 3.11.0
mysqlclient == 2.1.0
rsa == 4.9
pycryptodome == 3.17
qrcode